u-net-voc2012/requirements.txt
numpy==1.21.0
torch==1.9.0
torchvision==0.10.0
Pillow==8.2.0
matplotlib==3.4.3
scikit-learn==0.24.2
opencv-python==4.5.3.20210927
tqdm==4.62.3
PyYAML==5.4.1
jupyter==1.0.0